refactor: keep crawling order logic in views
